pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "flake8>=7.2.0",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "ruff>=0.11.5",
    "validators>=0.34.0",
]

[dependency-groups]
dev = [
    "flake8>=7.2.0",
]
